+++ server/src/controllers/MessageController.ts
@@ -1,0 +1,16 @@
+import MessageModel from "../models/MessageModel";
+
+class MessageController {
+    model: MessageModel
+
+    constructor(){
+        this.model = new MessageModel();
+    }
+
+    async getAllMessages() {
+        return await this.model.read();
+    }
+}
+
+
+export default MessageController;