+++ server/src/models/IModel.ts
@@ -1,6 +1,7 @@
 import { QueryResult } from "pg";
 
 interface IModel {
+    data(result: Promise<QueryResult<any>>): Promise<{}[]>;
     insert(...args: any[]): Promise<QueryResult<any>>;
     read(limit?: number, offset?: number): Promise<QueryResult<any>>;
     get(id: number): Promise<QueryResult<any>>;