+++ server/src/models/crud/CrudOrm.ts
@@ -1,0 +1,11 @@
+import connection from "../../connection";
+
+class CrudORM {
+    protected id: number ;
+    protected table: string;
+
+    
+    constructor(){
+
+    }
+}